backend/requirements.txt
Django>=5.1,<6
django-cors-headers>=4,<5
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
drf-spectacular>=0.27,<1
psycopg2-binary>=2.9,<3
cryptography>=42,<45
requests>=2.31,<3
polyline>=2,<3
reportlab>=4,<5
# Pin WeasyPrint to a Python 3.13-compatible release
weasyprint==61.2
